checkpointing

+++ src/stan/lang/ast/type/block_array_type.hpp
@@ -46,7 +46,7 @@ namespace stan {
        * Returns number of array dimensions.
        */
 
-      size_t dims() const;
+      int dims() const;
 
       /**
        * Returns the length of this array.
