+++ src/stan/lang/ast/type/local_array_type_def.hpp
@@ -13,7 +13,7 @@ namespace stan {
                                        const expression& len)
     : element_type_(el_type), array_len_(len) { }
 
-    size_t local_array_type::dims() const {
+    int local_array_type::dims() const {
       size_t total = 1;
       local_var_type cur_type(element_type_);
       while (cur_type.is_array_type()) {
